+++ src/cookbook/ScottPlot.Cookbook/Recipes/Quickstart.cs
